completed backend

+++ controllers/ContactControllers.js
@@ -1,14 +1,17 @@
-
 const asyncHandler = require("express-async-handler") // it removes the try catch block and detects error by itself
 const Contact = require("../models/contactModel")
 
-// get all contacts
+// @desc get all contacts
+// @Route /api/contact
+// @access private
 const getContacts = asyncHandler( async (req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({user_id : req.user.id})
     res.status(200).json(contacts)
 })
 
-// get a single contact
+// @desc get a contact
+// @Route /api/contact/:id
+// @access private
 const getContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.findById(req.params.id)
     if(!contact){
@@ -18,7 +21,9 @@ const getContact = asyncHandler(async (req, res)=>{
     res.status(200).json(contact)
 })
 
-//create a new contact
+// @desc create a contact
+// @Route /api/contact
+// @access private
 const createContacts = asyncHandler(async (req, res)=>{
     console.log("creating a user: ", req.body)
     const {email, name, phone} = req.body
@@ -30,12 +35,15 @@ const createContacts = asyncHandler(async (req, res)=>{
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     })
     res.status(201).json(contact)
 })
 
-// update a contact
+// @desc update a contact
+// @Route /api/contact/:id 
+// @access private
 const updateContact = asyncHandler(async (req, res)=>{
     const contact = await Contact.findById(req.params.id)
     if(!contact){
@@ -44,24 +52,35 @@ const updateContact = asyncHandler(async (req, res)=>{
         throw new Error("contact  not found")
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("You don't have permission to update")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body,{
         new: true
     })
     res.status(200).json(updatedContact)
 })
 
-// delete a contact
+// @desc delete a contact
+// @Route /api/contact/:id
+// @access private
 const deleteContact = asyncHandler(async (req, res)=>{
-    // const contact = await Contact.findById(req.params.id)
+    const contact = await Contact.findById(req.params.id)
     
-    // if(!contact){
-    //     console.log("inside not statement")
-    //     res.status(404);
-    //     throw new Error("contact  not found")
-    // }
+    if(!contact){
+        console.log("inside not statement")
+        res.status(404);
+        throw new Error("contact  not found")
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403)
+        throw new Error("You don't have permission to delete")
+    }
 
     const result  = await Contact.deleteOne({ _id: req.params.id })
-    res.status(200).json(result)
+    res.status(200).json(contact)
 })
 
 // exporting the controlls
